removes unused transformer forms, adds d.ts. files to js forms
