--- frontend/apps/web/app/(mgmt)/[account]/transformers/Sheetforms/GenerateCategoricalForm.tsx
@@ -1,79 +1,0 @@
-'use client';
-import { Button } from '@/components/ui/button';
-import {
-  Form,
-  FormControl,
-  FormDescription,
-  FormField,
-  FormItem,
-  FormLabel,
-} from '@/components/ui/form';
-import { Input } from '@/components/ui/input';
-import { yupResolver } from '@hookform/resolvers/yup';
-import { GenerateCategorical } from '@neosync/sdk';
-import { ReactElement } from 'react';
-import { useForm } from 'react-hook-form';
-import { TRANSFORMER_SCHEMA_CONFIGS } from '../../new/transformer/schema';
-import { TransformerFormProps } from './util';
-interface Props extends TransformerFormProps<GenerateCategorical> {}
-
-export default function GenerateCategoricalForm(props: Props): ReactElement {
-  const { existingConfig, onSubmit, isReadonly } = props;
-
-  const form = useForm({
-    mode: 'onChange',
-    resolver: yupResolver(TRANSFORMER_SCHEMA_CONFIGS.generateCategoricalConfig),
-    defaultValues: {
-      categories: existingConfig?.categories ?? '',
-    },
-  });
-
-  return (
-    <div className="flex flex-col w-full space-y-4 pt-4">
-      <Form {...form}>
-        <FormField
-          control={form.control}
-          name={`categories`}
-          render={({ field }) => (
-            <FormItem className="flex flex-row items-center justify-between rounded-lg border p-3 shadow-sm">
-              <div className="space-y-0.5">
-                <FormLabel>Categories</FormLabel>
-                <FormDescription>
-                  Provide a list of comma-separated string values that you want
-                  to randomly select from.
-                </FormDescription>
-              </div>
-              <FormControl>
-                <div className="max-w-[180px]">
-                  <Input
-                    {...field}
-                    type="string"
-                    className="max-w-[180px]"
-                    disabled={isReadonly}
-                  />
-                </div>
-              </FormControl>
-            </FormItem>
-          )}
-        />
-        <div className="flex justify-end">
-          <Button
-            type="button"
-            disabled={isReadonly}
-            onClick={(e) => {
-              form.handleSubmit((values) => {
-                onSubmit(
-                  new GenerateCategorical({
-                    ...values,
-                  })
-                );
-              })(e);
-            }}
-          >
-            Save
-          </Button>
-        </div>
-      </Form>
-    </div>
-  );
-}
