--- frontend/apps/web/app/(mgmt)/[account]/transformers/Sheetforms/GenerateGenderForm.tsx
@@ -1,72 +1,0 @@
-'use client';
-import { Button } from '@/components/ui/button';
-import {
-  Form,
-  FormControl,
-  FormDescription,
-  FormField,
-  FormItem,
-  FormLabel,
-} from '@/components/ui/form';
-import { Switch } from '@/components/ui/switch';
-import { yupResolver } from '@hookform/resolvers/yup';
-import { GenerateGender } from '@neosync/sdk';
-import { ReactElement } from 'react';
-import { useForm } from 'react-hook-form';
-import { TRANSFORMER_SCHEMA_CONFIGS } from '../../new/transformer/schema';
-import { TransformerFormProps } from './util';
-interface Props extends TransformerFormProps<GenerateGender> {}
-
-export default function GenerateGenderForm(props: Props): ReactElement {
-  const { existingConfig, onSubmit, isReadonly } = props;
-
-  const form = useForm({
-    mode: 'onChange',
-    resolver: yupResolver(TRANSFORMER_SCHEMA_CONFIGS.generateGenderConfig),
-    defaultValues: {
-      abbreviate: existingConfig?.abbreviate ?? false,
-    },
-  });
-
-  return (
-    <div className="flex flex-col w-full space-y-4 pt-4">
-      <Form {...form}>
-        <FormField
-          control={form.control}
-          name={`abbreviate`}
-          render={({ field }) => (
-            <FormItem className="flex flex-row items-center justify-between rounded-lg border p-3 shadow-sm">
-              <div className="space-y-0.5">
-                <FormLabel>Abbreviate</FormLabel>
-                <FormDescription>
-                  Abbreviate the gender to a single character. For example,
-                  female would be returned as f.
-                </FormDescription>
-              </div>
-              <FormControl>
-                <Switch
-                  disabled={isReadonly}
-                  checked={field.value}
-                  onCheckedChange={field.onChange}
-                />
-              </FormControl>
-            </FormItem>
-          )}
-        />
-        <div className="flex justify-end">
-          <Button
-            type="button"
-            disabled={isReadonly}
-            onClick={(e) => {
-              form.handleSubmit((values) => {
-                onSubmit(new GenerateGender({ ...values }));
-              })(e);
-            }}
-          >
-            Save
-          </Button>
-        </div>
-      </Form>
-    </div>
-  );
-}
