--- frontend/apps/web/app/(mgmt)/[account]/transformers/Sheetforms/TransformInt64PhoneForm.tsx
@@ -1,80 +1,0 @@
-'use client';
-import { Button } from '@/components/ui/button';
-import {
-  Form,
-  FormControl,
-  FormDescription,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormMessage,
-} from '@/components/ui/form';
-import { Switch } from '@/components/ui/switch';
-import { yupResolver } from '@hookform/resolvers/yup';
-import { TransformInt64PhoneNumber } from '@neosync/sdk';
-import { ReactElement } from 'react';
-import { useForm } from 'react-hook-form';
-import { TRANSFORMER_SCHEMA_CONFIGS } from '../../new/transformer/schema';
-import { TransformerFormProps } from './util';
-interface Props extends TransformerFormProps<TransformInt64PhoneNumber> {}
-
-export default function TransformInt64PhoneForm(props: Props): ReactElement {
-  const { existingConfig, onSubmit, isReadonly } = props;
-
-  const form = useForm({
-    mode: 'onChange',
-    resolver: yupResolver(
-      TRANSFORMER_SCHEMA_CONFIGS.transformInt64PhoneNumberConfig
-    ),
-    defaultValues: {
-      preserveLength: existingConfig?.preserveLength ?? false,
-    },
-  });
-
-  return (
-    <div className="flex flex-col w-full space-y-4 pt-4">
-      <Form {...form}>
-        <FormField
-          name={`preserveLength`}
-          render={({ field }) => (
-            <FormItem className="flex flex-row items-center justify-between rounded-lg border p-3 shadow-sm">
-              <div className="space-y-0.5">
-                <FormLabel>Preserve Length</FormLabel>
-                <FormDescription className="w-[90%]">
-                  Set the length of the output phone number to be the same as
-                  the input
-                </FormDescription>
-              </div>
-              <FormControl>
-                <Switch
-                  name={field.name}
-                  disabled={isReadonly}
-                  checked={field.value}
-                  onCheckedChange={field.onChange}
-                />
-              </FormControl>
-              <FormMessage />
-            </FormItem>
-          )}
-        />
-        <div className="flex justify-end">
-          <Button
-            type="button"
-            disabled={isReadonly}
-            onClick={(e) => {
-              form.handleSubmit((values) => {
-                onSubmit(
-                  new TransformInt64PhoneNumber({
-                    ...values,
-                  })
-                );
-              })(e);
-            }}
-          >
-            Save
-          </Button>
-        </div>
-      </Form>
-    </div>
-  );
-}
